Vulkan 0.5: Reorder files and code
